Split UserDB out of Spar.Scim
